+++ models/saes/matryoshka_sae.py
@@ -260,7 +260,7 @@ class MatryoshkaSAE(BaseSAE):
         
         # Reshape back to original shape
         acts_topk = acts_topk_flat.reshape(acts.shape)
-        mask = (acts_topk > 0).float()
+        mask = (acts_topk > 0).to(acts.dtype)
         
         return acts_topk, mask
 
@@ -295,7 +295,7 @@ class MatryoshkaSAE(BaseSAE):
         """
         # Convert threshold to acts dtype for comparison
         threshold = self.running_threshold.to(acts.dtype)
-        mask = (acts > threshold).float()
+        mask = (acts > threshold).to(acts.dtype)
         acts_sparse = acts * mask
         return acts_sparse, mask
